🔨 refactor: display mode 코드 분리
