+++ src/app/(user)/(service)/my-page/settings/page.tsx
@@ -8,8 +8,7 @@ import { useMutation } from "@tanstack/react-query";
 import client from "@/src/lib/api/client";
 import Swal from "sweetalert2";
 import LogoutButton from "@/src/features/auth/auth-common/components/logout-btn";
-import { DropdownThemeToggle } from "@/src/features/display-mode/display-mode-common/components/dark-mode-toggle";
-
+import { DropdownThemeToggle } from "@/src/features/display-mode/display-mode-common/components/display-mode-toggle";
 
 export default function Page() {
   const session = useSession();
